requirements.txt
# general modules
Django>4.2,<4.3
pillow
python-decouple
djangorestframework


# third party modules
markdown
django-filter
django-crispy-forms
crispy-bootstrap4
django-simple-captcha
drf-yasg[validation]
drf-spectacular[sidecar]
djangorestframework-simplejwt
django-mail-templated

# deployment modules

# refactoring and test
black
flake8